requirements.txt
django
requests
gunicorn
django-dotenv
psycopg2-binary
django-storages
boto3
